Add a custom constraint
This is  a custom constraint created by Simon Jeng in OpenSim Moco. This constraint is of MocoPathConstraint class, and constrains the relative angle between two frames. The constraint was developed for two-dimensional model in the sagittal plane (the plane perpendicular to the z axis, exactly), because 2D is enough in my research. If this constraint is needed in Moco, we can improve it to adapt three-dimensional models.

This is the first time for me to pull requests in Github. If more information is required, pleas let me know.

Thanks,
Simon Jeng